+++ redash/models/users.py
@@ -1,0 +1,448 @@
+import hashlib
+import itertools
+import logging
+from functools import reduce
+from operator import or_
+
+import walrus
+from flask import current_app as app, url_for, request_started
+from flask_login import current_user, AnonymousUserMixin, UserMixin
+from passlib.apps import custom_app_context as pwd_context
+from six import python_2_unicode_compatible, string_types, text_type
+from sqlalchemy.exc import DBAPIError
+from sqlalchemy.dialects import postgresql
+
+from sqlalchemy_utils import EmailType
+from sqlalchemy_utils.models import generic_repr
+
+from redash import redis_connection
+from redash.utils import generate_token, utcnow
+
+from .base import db, Column, GFKBase
+from .mixins import TimestampMixin, BelongsToOrgMixin
+from .types import json_cast_property, MutableDict, MutableList, UTCDateTimeField
+
+logger = logging.getLogger(__name__)
+
+
+class UserDetail(walrus.Model):
+    """
+    A walrus data model to store some user data to Redis to be
+    synced to Postgres asynchronously.
+    """
+    __database__ = redis_connection
+    __namespace__ = 'redash.user.details'
+
+    user_id = walrus.IntegerField(index=True)
+    updated_at = UTCDateTimeField(index=True, default=utcnow)
+
+    @classmethod
+    def update(cls, user_id):
+        """
+        Update the user details hash using the given redis
+        pipeline, user id, optional redis id and optional user
+        details.
+
+        The fields uid, rid and updated (timestamp) are
+        enforced and can't be overwritten.
+        """
+        # try getting the user detail with the given user ID
+        # or create one if it doesn't exist yet (e.g. when key was purged)
+        try:
+            user_detail = cls.get(cls.user_id == user_id)
+            # update the timestamp with the current time
+            user_detail.updated_at = utcnow()
+            # save to Redis
+            user_detail.save()
+        except ValueError:
+            user_detail = cls.create(
+                user_id=user_id,
+                updated_at=utcnow(),
+            )
+        return user_detail
+
+    @classmethod
+    def sync(cls, chunksize=1000):
+        """
+        Syncs user details to Postgres (to the JSON field User.details).
+        """
+        to_sync = {}
+        try:
+            for user_detail in cls.all():
+                to_sync[user_detail.user_id] = user_detail
+
+            user_ids = list(to_sync.keys())
+            if not user_ids:
+                return
+            logger.info(
+                'syncing users: %s',
+                ', '.join([str(uid) for uid in user_ids])
+            )
+            # get all SQLA users that need to be updated
+            users = User.query.filter(User.id.in_(user_ids))
+            for i, user in enumerate(users):
+                update = to_sync[user.id]
+                user.active_at = update.updated_at
+                # flush changes to the database after a certain
+                # number of items and extend the list of keys to
+                # stop sync in case of exceptions
+                if i % chunksize == 0:
+                    db.session.flush()
+            db.session.commit()
+        except DBAPIError:
+            # reset list of keys to stop sync
+            pass
+        finally:
+            user_ids = [str(user_id) for user_id in to_sync.keys()]
+            if user_ids:
+                logger.info(
+                    'Deleting temporary user details for users %s',
+                    ', '.join(user_ids)
+                )
+                delete_query = [
+                    UserDetail.user_id == str(user_id)
+                    for user_id in user_ids
+                ]
+                UserDetail.query_delete(reduce(or_, delete_query))
+
+
+def update_user_detail(sender, *args, **kwargs):
+    """
+    Used as a Flask request_started signal callback that adds
+    the current user's details to Redis
+    """
+    if (
+        current_user.get_id() and
+        current_user.is_authenticated and
+        not current_user.is_api_user()
+    ):
+        UserDetail.update(current_user.id)
+
+
+def init_app(app):
+    """
+    A Flask extension to keep user details updates in Redis and
+    sync it periodically to the database (User.details).
+    """
+    request_started.connect(update_user_detail, app)
+
+
+class PermissionsCheckMixin(object):
+    def has_permission(self, permission):
+        return self.has_permissions((permission,))
+
+    def has_permissions(self, permissions):
+        has_permissions = reduce(lambda a, b: a and b,
+                                 map(lambda permission: permission in self.permissions,
+                                     permissions),
+                                 True)
+
+        return has_permissions
+
+
+@python_2_unicode_compatible
+@generic_repr('id', 'name', 'email')
+class User(TimestampMixin, db.Model, BelongsToOrgMixin, UserMixin, PermissionsCheckMixin):
+    id = Column(db.Integer, primary_key=True)
+    org_id = Column(db.Integer, db.ForeignKey('organizations.id'))
+    org = db.relationship("Organization", backref=db.backref("users", lazy="dynamic"))
+    name = Column(db.String(320))
+    email = Column(EmailType)
+    _profile_image_url = Column('profile_image_url', db.String(320), nullable=True)
+    password_hash = Column(db.String(128), nullable=True)
+    # XXX replace with association table
+    group_ids = Column('groups', MutableList.as_mutable(postgresql.ARRAY(db.Integer)), nullable=True)
+    api_key = Column(db.String(40),
+                     default=lambda: generate_token(40),
+                     unique=True)
+
+    disabled_at = Column(db.DateTime(True), default=None, nullable=True)
+    details = Column(MutableDict.as_mutable(postgresql.JSON), nullable=True,
+                     server_default='{}', default={})
+    active_at = json_cast_property(db.DateTime(True), 'details', 'active_at',
+                                   default=None)
+
+    __tablename__ = 'users'
+    __table_args__ = (
+        db.Index('users_org_id_email', 'org_id', 'email', unique=True),
+    )
+
+    def __str__(self):
+        return u'%s (%s)' % (self.name, self.email)
+
+    def __init__(self, *args, **kwargs):
+        if kwargs.get('email') is not None:
+            kwargs['email'] = kwargs['email'].lower()
+        super(User, self).__init__(*args, **kwargs)
+
+    @property
+    def is_disabled(self):
+        return self.disabled_at is not None
+
+    def disable(self):
+        self.disabled_at = db.func.now()
+
+    def enable(self):
+        self.disabled_at = None
+
+    def to_dict(self, with_api_key=False):
+        profile_image_url = self.profile_image_url
+        if self.is_disabled:
+            assets = app.extensions['webpack']['assets'] or {}
+            path = 'images/avatar.svg'
+            profile_image_url = url_for('static', filename=assets.get(path, path))
+
+        d = {
+            'id': self.id,
+            'name': self.name,
+            'email': self.email,
+            'profile_image_url': profile_image_url,
+            'groups': self.group_ids,
+            'updated_at': self.updated_at,
+            'created_at': self.created_at,
+            'disabled_at': self.disabled_at,
+            'is_disabled': self.is_disabled,
+            'active_at': self.active_at,
+        }
+
+        if self.password_hash is None:
+            d['auth_type'] = 'external'
+        else:
+            d['auth_type'] = 'password'
+
+        if with_api_key:
+            d['api_key'] = self.api_key
+
+        return d
+
+    def is_api_user(self):
+        return False
+
+    @property
+    def profile_image_url(self):
+        if self._profile_image_url is not None:
+            return self._profile_image_url
+
+        email_md5 = hashlib.md5(self.email.lower()).hexdigest()
+        return "https://www.gravatar.com/avatar/{}?s=40&d=identicon".format(email_md5)
+
+    @property
+    def permissions(self):
+        # TODO: this should be cached.
+        return list(itertools.chain(*[g.permissions for g in
+                                      Group.query.filter(Group.id.in_(self.group_ids))]))
+
+    @classmethod
+    def get_by_org(cls, org):
+        return cls.query.filter(cls.org == org)
+
+    @classmethod
+    def get_by_email_and_org(cls, email, org):
+        return cls.get_by_org(org).filter(cls.email == email).one()
+
+    @classmethod
+    def get_by_api_key_and_org(cls, api_key, org):
+        return cls.get_by_org(org).filter(cls.api_key == api_key).one()
+
+    @classmethod
+    def all(cls, org):
+        return cls.get_by_org(org).filter(cls.disabled_at.is_(None))
+
+    @classmethod
+    def search(cls, base_query, term):
+        term = u'%{}%'.format(term)
+        search_filter = or_(cls.name.ilike(term), cls.email.like(term))
+
+        return base_query.filter(search_filter)
+
+    @classmethod
+    def all_disabled(cls, org):
+        return cls.get_by_org(org).filter(cls.disabled_at.isnot(None))
+
+    @classmethod
+    def find_by_email(cls, email):
+        return cls.query.filter(cls.email == email)
+
+    def hash_password(self, password):
+        self.password_hash = pwd_context.encrypt(password)
+
+    def verify_password(self, password):
+        return self.password_hash and pwd_context.verify(password, self.password_hash)
+
+    def update_group_assignments(self, group_names):
+        groups = Group.find_by_name(self.org, group_names)
+        groups.append(self.org.default_group)
+        self.group_ids = [g.id for g in groups]
+        db.session.add(self)
+        db.session.commit()
+
+    def has_access(self, obj, access_type):
+        return AccessPermission.exists(obj, access_type, grantee=self)
+
+    def get_id(self):
+        identity = hashlib.md5(
+            "{},{}".format(self.email, self.password_hash)
+        ).hexdigest()
+        return u"{0}-{1}".format(self.id, identity)
+
+
+@python_2_unicode_compatible
+@generic_repr('id', 'name', 'type', 'org_id')
+class Group(db.Model, BelongsToOrgMixin):
+    DEFAULT_PERMISSIONS = ['create_dashboard', 'create_query', 'edit_dashboard', 'edit_query',
+                           'view_query', 'view_source', 'execute_query', 'list_users', 'schedule_query',
+                           'list_dashboards', 'list_alerts', 'list_data_sources']
+
+    BUILTIN_GROUP = 'builtin'
+    REGULAR_GROUP = 'regular'
+
+    id = Column(db.Integer, primary_key=True)
+    data_sources = db.relationship("DataSourceGroup", back_populates="group",
+                                   cascade="all")
+    org_id = Column(db.Integer, db.ForeignKey('organizations.id'))
+    org = db.relationship("Organization", back_populates="groups")
+    type = Column(db.String(255), default=REGULAR_GROUP)
+    name = Column(db.String(100))
+    permissions = Column(postgresql.ARRAY(db.String(255)),
+                         default=DEFAULT_PERMISSIONS)
+    created_at = Column(db.DateTime(True), default=db.func.now())
+
+    __tablename__ = 'groups'
+
+    def __str__(self):
+        return text_type(self.id)
+
+    def to_dict(self):
+        return {
+            'id': self.id,
+            'name': self.name,
+            'permissions': self.permissions,
+            'type': self.type,
+            'created_at': self.created_at
+        }
+
+    @classmethod
+    def all(cls, org):
+        return cls.query.filter(cls.org == org)
+
+    @classmethod
+    def members(cls, group_id):
+        return User.query.filter(User.group_ids.any(group_id))
+
+    @classmethod
+    def find_by_name(cls, org, group_names):
+        result = cls.query.filter(cls.org == org, cls.name.in_(group_names))
+        return list(result)
+
+
+@generic_repr('id', 'object_type', 'object_id', 'access_type', 'grantor_id', 'grantee_id')
+class AccessPermission(GFKBase, db.Model):
+    id = Column(db.Integer, primary_key=True)
+    # 'object' defined in GFKBase
+    access_type = Column(db.String(255))
+    grantor_id = Column(db.Integer, db.ForeignKey("users.id"))
+    grantor = db.relationship(User, backref='grantor', foreign_keys=[grantor_id])
+    grantee_id = Column(db.Integer, db.ForeignKey("users.id"))
+    grantee = db.relationship(User, backref='grantee', foreign_keys=[grantee_id])
+
+    __tablename__ = 'access_permissions'
+
+    @classmethod
+    def grant(cls, obj, access_type, grantee, grantor):
+        grant = cls.query.filter(cls.object_type == obj.__tablename__,
+                                 cls.object_id == obj.id,
+                                 cls.access_type == access_type,
+                                 cls.grantee == grantee,
+                                 cls.grantor == grantor).one_or_none()
+
+        if not grant:
+            grant = cls(object_type=obj.__tablename__,
+                        object_id=obj.id,
+                        access_type=access_type,
+                        grantee=grantee,
+                        grantor=grantor)
+            db.session.add(grant)
+
+        return grant
+
+    @classmethod
+    def revoke(cls, obj, grantee, access_type=None):
+        permissions = cls._query(obj, access_type, grantee)
+        return permissions.delete()
+
+    @classmethod
+    def find(cls, obj, access_type=None, grantee=None, grantor=None):
+        return cls._query(obj, access_type, grantee, grantor)
+
+    @classmethod
+    def exists(cls, obj, access_type, grantee):
+        return cls.find(obj, access_type, grantee).count() > 0
+
+    @classmethod
+    def _query(cls, obj, access_type=None, grantee=None, grantor=None):
+        q = cls.query.filter(cls.object_id == obj.id,
+                             cls.object_type == obj.__tablename__)
+
+        if access_type:
+            q = q.filter(AccessPermission.access_type == access_type)
+
+        if grantee:
+            q = q.filter(AccessPermission.grantee == grantee)
+
+        if grantor:
+            q = q.filter(AccessPermission.grantor == grantor)
+
+        return q
+
+    def to_dict(self):
+        d = {
+            'id': self.id,
+            'object_id': self.object_id,
+            'object_type': self.object_type,
+            'access_type': self.access_type,
+            'grantor': self.grantor_id,
+            'grantee': self.grantee_id
+        }
+        return d
+
+
+class AnonymousUser(AnonymousUserMixin, PermissionsCheckMixin):
+    @property
+    def permissions(self):
+        return []
+
+    def is_api_user(self):
+        return False
+
+
+class ApiUser(UserMixin, PermissionsCheckMixin):
+    def __init__(self, api_key, org, groups, name=None):
+        self.object = None
+        if isinstance(api_key, string_types):
+            self.id = api_key
+            self.name = name
+        else:
+            self.id = api_key.api_key
+            self.name = "ApiKey: {}".format(api_key.id)
+            self.object = api_key.object
+        self.group_ids = groups
+        self.org = org
+
+    def __repr__(self):
+        return u"<{}>".format(self.name)
+
+    def is_api_user(self):
+        return True
+
+    @property
+    def org_id(self):
+        if not self.org:
+            return None
+        return self.org.id
+
+    @property
+    def permissions(self):
+        return ['view_query']
+
+    def has_access(self, obj, access_type):
+        return False
