+++ redash/tasks/__init__.py
@@ -1,3 +1,3 @@
-from .general import record_event, version_check, send_mail
+from .general import record_event, version_check, send_mail, sync_user_details
 from .queries import QueryTask, refresh_queries, refresh_schemas, cleanup_tasks, cleanup_query_results, execute_query
 from .alerts import check_alerts_for_query
