+++ requirements.txt
@@ -10,7 +10,7 @@ Flask-Admin==1.5.2
 Flask-RESTful==0.3.5
 Flask-Login==0.4.0
 Flask-OAuthLib==0.9.3
-Flask-SQLAlchemy==2.1
+Flask-SQLAlchemy==2.3.2
 Flask-Migrate==2.0.1
 flask-mail==0.9.1
 flask-sslify==0.1.5
@@ -19,19 +19,21 @@ passlib==1.6.2
 aniso8601==1.1.0
 blinker==1.3
 psycopg2==2.7.3.2
-python-dateutil==2.4.2
+python-dateutil==2.7.3
 pytz==2016.7
 PyYAML==3.12
-redis==2.10.5
+redis==3.0.1
+walrus==0.7.1
 requests==2.21.0
 six==1.11.0
-SQLAlchemy==1.2.7
+SQLAlchemy==1.2.12
 SQLAlchemy-Searchable==0.10.6
 SQLAlchemy-Utils>=0.29.0
 sqlparse==0.2.4
 statsd==2.1.2
 gunicorn==19.7.1
 celery==4.2.1
+kombu==4.2.2.post1
 jsonschema==2.4.0
 RestrictedPython==3.6.0
 pysaml2==4.5.0
