+++ setup.cfg
@@ -1,2 +1,7 @@
 [pep8]
 max-line-length = 120
+
+[flake8]
+ignore = E501
+exclude = .git
+max-complexity = 10
