+++ tests/handlers/test_visualizations.py
@@ -27,7 +27,7 @@ class VisualizationResourceTest(BaseTestCase):
         rv = self.make_request('delete', '/api/visualizations/{}'.format(visualization.id))
 
         self.assertEquals(rv.status_code, 200)
-        self.assertEquals(models.db.session.query(models.Visualization).count(), 0)
+        self.assertEquals(models.Visualization.query.count(), 0)
 
     def test_update_visualization(self):
         visualization = self.factory.create_visualization()
